requirements.txt
# ETL Pipeline for PDF Document Processing
# Install with: pip install -r requirements.txt

# PDF Processing
pymupdf>=1.24.0
pymupdf4llm>=0.0.5

# Text Processing
langchain>=0.2.0
langchain-text-splitters>=0.2.0
tiktoken>=0.5.0

# Embeddings
sentence-transformers>=2.2.0
openai>=1.0.0

# Vector Database (Local) - LanceDB with native hybrid search
lancedb>=0.4.0
pyarrow>=14.0.0

# Database
# supabase>=2.0.0  # Uncomment for production

# Image Processing
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pydantic>=2.0.0
pyyaml>=6.0.0

# Development
pytest>=7.0.0
pytest-cov>=4.0.0
